requirements.txt
pip~=25.0
Django~=5.1
asgiref~=3.8.1
sqlparse~=0.5.1
djangorestframework~=3.15.2
django-filter~=25.1